Get Elixhauser Codes

Retrieve copy of internal lookup tables for the ICD codes used in assessing Elixhauser comorbidities.

get_elixhauser_codes()

Value

A data.frame with the following columns:

icdv: Integer vector indicating if the code is from ICD-9 or ICD-10

dx: Integer vector. 1 if the code is a diagnostic, (ICD-9-CM, ICD-10-CM, WHO, CDC Mortality), or 0 if the code is procedural (ICD-9-PCS, ICD-10-PCS)

full_code: Character vector with the ICD code and any relevant decimal point

code: Character vector with the compact ICD code omitting any relevant decimal point

condition: Character vector of the conditions

elixhauser_<variant>: Integer vector indicating if the code is part of the <variant> of the Elixhauser comorbidities.

See also

Examples

head(get_elixhauser_codes())
#>   icdv dx full_code  code condition elixhauser_ahrq_web
#> 1    9  1       042   042      AIDS                   1
#> 2    9  1    070.22 07022     LIVER                   1
#> 3    9  1    070.23 07023     LIVER                   1
#> 4    9  1    070.32 07032     LIVER                   1
#> 5    9  1    070.33 07033     LIVER                   1
#> 6    9  1    070.44 07044     LIVER                   1
#>   elixhauser_elixhauser1988 elixhauser_quan2005 elixhauser_ahrq2022
#> 1                         1                   1                  NA
#> 2                         0                   1                  NA
#> 3                         0                   1                  NA
#> 4                         1                   1                  NA
#> 5                         1                   1                  NA
#> 6                         0                   1                  NA
#>   elixhauser_ahrq2023 elixhauser_ahrq2024 elixhauser_ahrq2025
#> 1                  NA                  NA                  NA
#> 2                  NA                  NA                  NA
#> 3                  NA                  NA                  NA
#> 4                  NA                  NA                  NA
#> 5                  NA                  NA                  NA
#> 6                  NA                  NA                  NA
str(get_elixhauser_codes())
#> 'data.frame':	10679 obs. of  12 variables:
#>  $ icdv                     : int  9 9 9 9 9 9 9 9 9 9 ...
#>  $ dx                       : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ full_code                : chr  "042" "070.22" "070.23" "070.32" ...
#>  $ code                     : chr  "042" "07022" "07023" "07032" ...
#>  $ condition                : chr  "AIDS" "LIVER" "LIVER" "LIVER" ...
#>  $ elixhauser_ahrq_web      : int  1 1 1 1 1 1 1 0 0 0 ...
#>  $ elixhauser_elixhauser1988: int  1 0 0 1 1 0 1 0 0 0 ...
#>  $ elixhauser_quan2005      : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ elixhauser_ahrq2022      : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ elixhauser_ahrq2023      : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ elixhauser_ahrq2024      : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ elixhauser_ahrq2025      : int  NA NA NA NA NA NA NA NA NA NA ...