Get Elixhauser Present-on-Admission Requirements

Retrieve a copy of internal lookup table with details on which Elixhauser comorbidities do and do not require the associated ICD codes to be present-on-admission to be flagged.

Usage

get_elixhauser_poa()

Value

A data.frame with the following columns:

condition: Character vector of the conditions

desc: Character vector with a verbose description of the condition

poa_required: Integer indicators if the code needs to present on admission to be considered a comorbidity

elixhauser_<variant>: indicators for the Elixhauser <variant>

See also

Examples

head(get_elixhauser_poa())
#>    condition                                          desc poa_required
#> 1       AIDS           Acquired immune deficiency syndrome            0
#> 2    ALCOHOL                                 Alcohol abuse            0
#> 3    ANEMDEF Anemias due to other nutritional deficiencies            1
#> 4    ANEMDEF                            Deficiency anemias            1
#> 5 AUTOIMMUNE                         Autoimmune conditions            0
#> 6    BLDLOSS          Chronic blood loss (iron deficiency)            1
#>   elixhauser_ahrq2022 elixhauser_ahrq2023 elixhauser_ahrq2024
#> 1                   1                   1                   1
#> 2                   1                   1                   1
#> 3                  NA                   1                   1
#> 4                   1                  NA                  NA
#> 5                   1                   1                   1
#> 6                  NA                   1                   1
#>   elixhauser_ahrq2025
#> 1                   1
#> 2                   1
#> 3                   1
#> 4                  NA
#> 5                   1
#> 6                   1
str(get_elixhauser_poa())
#> 'data.frame':	43 obs. of  7 variables:
#>  $ condition          : chr  "AIDS" "ALCOHOL" "ANEMDEF" "ANEMDEF" ...
#>  $ desc               : chr  "Acquired immune deficiency syndrome" "Alcohol abuse" "Anemias due to other nutritional deficiencies" "Deficiency anemias" ...
#>  $ poa_required       : int  0 0 1 1 0 1 1 0 0 0 ...
#>  $ elixhauser_ahrq2022: int  1 1 NA 1 1 NA 1 1 1 1 ...
#>  $ elixhauser_ahrq2023: int  1 1 1 NA 1 1 NA 1 1 1 ...
#>  $ elixhauser_ahrq2024: int  1 1 1 NA 1 1 NA 1 1 1 ...
#>  $ elixhauser_ahrq2025: int  1 1 1 NA 1 1 NA 1 1 1 ...